lab 7,8

+++ src/main/java/com/example/demo/cache/Cache.java
@@ -1,24 +1,25 @@
 package com.example.demo.cache;
 
-import java.util.*;
-
 import com.example.demo.logic.SinIntegral;
 import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Component
 public class Cache {
 
-    private final Map<Integer, SinIntegral> cacheMap = new HashMap<>();
+    private final Map<Long, SinIntegral> cacheMap = new HashMap<>();
 
-    public void put(Integer key, SinIntegral value) {
+    public void put(Long key, SinIntegral value) {
         cacheMap.put(key, value);
     }
 
-    public SinIntegral get(Integer key) {
+    public SinIntegral get(Long key) {
         return cacheMap.get(key);
     }
 
-    public void remove(Integer key) {
+    public void remove(Long key) {
         cacheMap.remove(key);
     }
